+++ savalidation/__init__.py
@@ -58,10 +58,6 @@ class _ValidationHelper(object):
         self.entref = weakref.ref(entity)
         self.field_validators = defaultdict(list)
         self.chained_validators = []
-        self.before_flush_methods = []
-
-        self.init_fe_validators()
-        self.init_before_flush_methods()
 
     @property
     def entity(self):
@@ -95,28 +91,9 @@ class _ValidationHelper(object):
     def entity_linkers(self):
         return self.entity._sav_entity_linkers
 
-    def init_fe_validators(self):
-        for val_class, args, kwargs in self.entity_linkers:
-            # val_class should be a subclass of ValidatorBase
-            sav_val = val_class(self.entity.__class__, *args, **kwargs)
-
-            # chained FE validators from entity SAV validators
-            self.chained_validators.extend(sav_val.fe_entity_validators)
-
-            # field FE validators from field SAV validators
-            for field_name, fe_val in sav_val.fe_field_validators:
-                #print field_name, fe_val
-                self.field_validators[field_name].append(fe_val)
-
-    def init_before_flush_methods(self):
-        for attr_name, attr_obj in self.entity.__class__.__dict__.iteritems():
-            # test for a value, not just the presence of the attribute to avoid
-            # collecting methods that have been mocked and will therefore
-            # have all attributes.
-            if getattr(attr_obj, '_sav_before_flush', False) == 'yes':
-                # don't want to get into trouble w/ strong references, so only
-                # keep a copy of the name of the attribute
-                self.before_flush_methods.append(attr_name)
+    @property
+    def before_flush_methods(self):
+        return self.entity._sav_before_flush_methods
 
     def trigger_before_flush_methods(self):
         for mname in self.before_flush_methods:
@@ -129,58 +106,16 @@ class _ValidationHelper(object):
     def add_error(self, field_name, msg):
         self.errors[field_name].append(msg)
 
-    def validate(self, type):
-        if type == 'before_flush':
-            self.clear_errors()
-            self.trigger_before_flush_methods()
-
-        # if there were no validators setup, then just return
-        if not self.entity_linkers:
-            # make sure we return errors here since a before_flush() method
-            # could have added validation errors
-            return self.errors
-
-        fe_val_schema, fe_conv_schema = self.create_fe_schemas(type)
-        self.validate_fe_schema(fe_val_schema, False)
-        self.validate_fe_schema(fe_conv_schema, True)
-        return self.errors
-
-    def does_validator_apply(self, fieldname, fe_val, type):
-        def_on_none = fe_val._sav_defer_on_none
-        if not def_on_none:
-            return type == 'before_flush'
-        fvalue = self.entity.__dict__.get(fieldname, None)
-        if fvalue is not None:
-            return type == 'before_flush'
-        return type != 'before_flush'
-
-    def create_fe_schemas(self, type):
-        fe_val_schema = formencode.Schema(allow_extra_fields = True)
-        fe_conv_schema = formencode.Schema(allow_extra_fields = True)
-        for field, validators in self.field_validators.iteritems():
-            validators_to_apply = []
-            converters_to_apply = []
-            for v in validators:
-                if self.does_validator_apply(field, v, type):
-                    if v._sav_convert_flag:
-                        converters_to_apply.append(v)
-                    else:
-                        validators_to_apply.append(v)
-            if validators_to_apply:
-                fe_val_schema.add_field(field, formencode.compound.All(*validators_to_apply))
-            if converters_to_apply:
-                fe_conv_schema.add_field(field, formencode.compound.All(*converters_to_apply))
-        for fe_validator in self.chained_validators:
-            fe_schema.add_chained_validator(fe_validator)
-        return fe_val_schema, fe_conv_schema
-
     def validate_fe_schema(self, schema, flag_convert):
+        if not schema.fields:
+            # print 'return b/c schema has no fields'
+            return
         idict = {}
         for colname in self.entity._sav_column_names():
             if schema.fields.has_key(colname):
                 idict[colname] = getattr(self.entity, colname, None)
         try:
-            #print '-------------', idict, schema, flag_convert, self.entity
+            # print '-------------', idict, schema, flag_convert, self.entity
             processed = schema.to_python(idict, _FEState(self.entity))
             if flag_convert:
                 self.entity.__dict__.update(processed)
@@ -188,9 +123,21 @@ class _ValidationHelper(object):
         except formencode.Invalid, e:
             for field_name, msg in e.unpack_errors().iteritems():
                 self.add_error(field_name, msg)
+            return True
+        return False
+
+    def run_event_schemas(self, event):
+        has_error = False
+        if self.entity._sav_fe_schemas:
+            val_schema, conv_schema = self.entity._sav_fe_schemas[event]
+            if self.validate_fe_schema(val_schema, False):
+                has_error = True
+            if self.validate_fe_schema(conv_schema, True):
+                has_error = True
+        return has_error
 
 class ValidationMixin(object):
-    _sav_entity_linkers = ()
+    _sav_do_validation = True
 
     def _sav_initialize(self):
         self._sav = _ValidationHelper(self)
@@ -212,6 +159,77 @@ class ValidationMixin(object):
                      for name in self.sa_column_names() if name not in exclude])
         return data
 
+    @staticmethod
+    @sa.event.listens_for(saorm.mapper, 'mapper_configured')
+    def _sav_init_validation_class(mapper, cls):
+        """
+            This method is called when the class is finished initializing to
+            take all the validation info that was associated with the class
+            and turn it into formencode schemas that can be used later by the
+            instances to do validation.
+        """
+
+        # dont run for classes that aren't validators
+        if not getattr(cls, '_sav_do_validation', False):
+            return
+
+        # only want this to run once per class
+        if getattr(cls, '_sav_class_init_already_ran', False):
+            return
+
+        # make sure the attribute is there for classes which use the mixin
+        # but don't actually use any validators
+        cls._sav_class_init_already_ran = True
+        if not hasattr(cls, '_sav_entity_linkers'):
+            cls._sav_entity_linkers = ()
+        cls._sav_fe_schemas = {}
+        cls._sav_before_flush_methods = []
+
+        # gather all the fev_metas from all entity linkers into one place
+        all_fev_metas = []
+        for val_class, args, kwargs in cls._sav_entity_linkers:
+            # val_class should be a subclass of ValidatorBase
+            sav_val = val_class(cls, *args, **kwargs)
+            all_fev_metas.extend(sav_val.fev_metas)
+
+        # create the formencode schemas to validate with
+        cls._sav_fe_schemas['before_flush'] = \
+            cls._sav_create_fe_schema(all_fev_metas, 'before_flush', False), \
+            cls._sav_create_fe_schema(all_fev_metas, 'before_flush', True)
+
+        cls._sav_fe_schemas['before_exec'] = \
+            cls._sav_create_fe_schema(all_fev_metas, 'before_exec', False), \
+            cls._sav_create_fe_schema(all_fev_metas, 'before_exec', True)
+
+        # setup methods that have been decorated with the before_flush event
+        for attr_name, attr_obj in cls.__dict__.iteritems():
+            # test for a value, not just the presence of the attribute to avoid
+            # collecting methods that have been mocked and will therefore
+            # have all attributes.
+            if getattr(attr_obj, '_sav_before_flush', False) == 'yes':
+                # don't want to get into trouble w/ strong references, so only
+                # keep a copy of the name of the attribute
+                cls._sav_before_flush_methods.append(attr_name)
+
+    @classmethod
+    def _sav_create_fe_schema(cls, fev_metas, for_event, for_conversion):
+        schema = formencode.Schema(allow_extra_fields = True)
+        field_validators = defaultdict(list)
+        for fevm in fev_metas:
+            if fevm.event == for_event and fevm.is_converter == for_conversion:
+                field_validators[fevm.field_name].append(fevm.fev)
+        for fieldname, validators in field_validators.iteritems():
+            schema.add_field(fieldname, formencode.compound.All(*validators))
+        return schema
+
+    @classmethod
+    def _sav_validate(cls, instance, type):
+        if type == 'before_flush':
+            instance._sav.clear_errors()
+            instance._sav.trigger_before_flush_methods()
+
+        return instance._sav.run_event_schemas(type)
+
 class _EventHandler(object):
 
     @staticmethod
@@ -226,61 +244,52 @@ class _EventHandler(object):
         if hasattr(target, '_sav_initialize'):
             target._sav_initialize()
 
-    @classmethod
-    def do_validation(cls, insts_to_val, insts_with_err, type):
-        for instance in insts_to_val:
-            if instance in insts_with_err:
-                continue
+    @staticmethod
+    def handle_insert(mapper, connection, target):
+        if hasattr(target, '_sav_validate'):
+            errors = target._sav_validate(target, 'before_exec')
 
-            errors = instance._sav.validate(type)
-            if errors:
-                insts_with_err.append(instance)
+    @staticmethod
+    @sa.event.listens_for(saorm.mapper, 'before_insert')
+    @sa.event.listens_for(saorm.mapper, 'before_update')
+    def handle_before_exec(mapper, connection, target):
+        if not hasattr(target, '_sav_validate'):
+            return
+        sess = saorm.session.Session.object_session(target)
+        sess._sav_ent_exec_count -= 1
 
-    @classmethod
-    def restore_and_raise(cls, session, iwe, expunged):
-        # add the instances back in that got expunged
-        for instance in expunged:
-            session.add(instance)
-        if iwe:
-            raise ValidationError(iwe)
+        target._sav_validate(target, 'before_exec')
+
+        if sess._sav_ent_exec_count == 0:
+            ents_with_error = []
+            for ent in sess._sav_ents_to_validate:
+                if ent.validation_errors:
+                    ents_with_error.append(ent)
+            # print ents_with_error
+            if ents_with_error:
+                raise ValidationError(ents_with_error)
 
     @classmethod
     def before_flush(cls, session, flush_context, instances):
-        iwe = session._sav_insts_with_err = []
-        itv = session._sav_insts_to_val = []
+        ents_to_validate = session._sav_ents_to_validate = []
 
-        for inst in session.new:
-            if not isinstance(inst, ValidationMixin):
+        for ent in session.new:
+            if not hasattr(ent, '_sav_validate'):
                 continue
-            itv.append(inst)
+            ents_to_validate.append(ent)
 
-        for inst in session.dirty:
-            if not isinstance(inst, ValidationMixin):
+        for ent in session.dirty:
+            if not hasattr(ent, '_sav_validate'):
                 continue
+            if session.is_modified(ent):
+                ents_to_validate.append(ent)
 
-            if session.is_modified(inst):
-                itv.append(inst)
+        # save the number of instances so we know when to raise in the
+        # handle_before_exec() method above
+        session._sav_ent_exec_count = len(ents_to_validate)
 
-        cls.do_validation(itv, iwe, 'before_flush')
-
-        # expunge all instances with an error from the session so that they
-        # don't get flushed.  They will get added back into the session later
-        # so that the state is consistent.  We just want to prevent DB errors
-        # as much as possible.
-        for instance in iwe:
-            session.expunge(instance)
-
-        # if we have expunged all the instances
-        if not session.new and not session.dirty and not session.deleted:
-            cls.restore_and_raise(session, iwe, iwe)
-
-    @classmethod
-    def after_flush(cls, session, flush_context):
-        iwe = session._sav_insts_with_err
-        expunged = list(iwe)
-        itv = session._sav_insts_to_val
-        cls.do_validation(itv, iwe, 'after_flush')
-        cls.restore_and_raise(session, iwe, expunged)
+        for ent in ents_to_validate:
+            ent._sav_validate(ent, 'before_flush')
 
 # until this bug gets fixed & released:
 #   http://www.sqlalchemy.org/trac/ticket/2424
@@ -288,8 +297,6 @@ class _EventHandler(object):
 # is created.  If that is not the case, then call watch_session() with your
 # session object and the events will be registered correctly.
 sa.event.listen(saorm.Session, 'before_flush', _EventHandler.before_flush)
-sa.event.listen(saorm.Session, 'after_flush', _EventHandler.after_flush)
 
 def watch_session(sess):
     sa.event.listen(sess, 'before_flush', _EventHandler.before_flush)
-    sa.event.listen(sess, 'after_flush', _EventHandler.after_flush)